OT2 modifications for ADT workflow
The .json files constitute the custom labware definitions for the ADT workflow. The modifications and additional scripts have added the functions and commands needed to execute the workflow with an OT2 robot.
The Jupyter notebook was created to execute the protocol through the advanced control option in Opentrons. It includes additional parameters, such as the volume of aliquots extracted from each donor and receptor chamber, the volume of water used to dilute the aliquots in the sample well plate, total experimental time and frequency of sample collection. The notebook accommodates modifications of the parameters and experimental criteria, particularly for small aqueous dyes.
Additional software capabilities include the ability to accommodate increased or decreased diffusion rates,  arbitrary frequencies for sampling, and address limitations in the sampling hardware.