added docstrings to the ADT script, updated testing notebook, modified readme
Edited the OT2_Hcell_command script to include docstrings for all functions and added a "dye injection" capability to allow the robot to distribute the stock dye to each donor chamber before conducting iteration 0 of sample collection. Additionally, fixed the water volume tracking within the main protocol function to properly account for the number of H-cells being tested.
Made minor edits to the ADT_Hcell_Testing notebook (added dye injection inputs to the main function and added explanations).
Edited the README.md to match the content of the current repo and added explanations and instructions for implementation of the workflow.
Added the ADT Data Analysis folder, which contains sample notebooks for processing the data acquired from the ADT workflow. The walkthrough notebook contains explanations and an example set of data; the "blank" notebook contains the functions without the additional explanations. Added an "assets" folder with figures as visual aids for the hardware.